programming Session and Cookies

+++ route.py
@@ -1,8 +1,17 @@
-from flask import Flask,render_template, redirect, url_for, request
+from flask import Flask,render_template, redirect, url_for, request, session
 import  db_module
+from flask_login import LoginManager
 
 app = Flask(__name__)
 
+app.secret_key = 'YouWillNeverGuess'
+##
+## for logging out session.pop('logged_in')
+##
+def check_logged_in() -> bool:
+    if 'logged_in' in session:
+        return True
+    return False
 
 @app.route("/")
 def entry():
@@ -21,8 +30,13 @@ def login():
             return str(error)
             #error = data
         elif exist and username =='admin' :
+
+            session['logged_in'] = True
+            session['username'] = username
             return redirect(url_for('admin_console'))
         else:
+            session['logged_in'] = True
+            session['username'] = username
             return redirect(url_for('restaurant'))
     return render_template('login.html', error=error)
 
@@ -53,23 +67,29 @@ def registration():
 @app.route('/restaurants', methods=['GET'])
 def restaurant():
     error = None
-    if request.method == 'GET':
-        data = db_module.get_rest()
-    return render_template('restaurants.html', restaurants=data)
+    
+    if request.method == 'GET' and check_logged_in():
+        data = db_module.get_rest()    
+        print (session['username'])
+        return render_template('restaurants.html', restaurants=data)
+    else:
+        return redirect(url_for('login'))
 
 
 
 @app.route('/admin_console', methods=['GET','POST'])
 def admin_console():
     error = None
-    if request.method == 'GET':
+    if request.method == 'GET' and check_logged_in() and session['username'] == 'admin':
         #select = request.form.get('actions')
         #print (select)
         rest_ = db_module.get_rest()
-        category_ = db_module.get_category()
+        #category_ = db_module.get_category()
         orders_ = db_module.get_order()
-        return render_template('console.html', rest=rest_,orders=orders_,category=category_)
-    else:
+        data_ = db_module.get_category()
+        print (data_)
+        return render_template('console.html', rest=rest_,data=data_,order_view=orders_)
+    elif request.method == 'POST' and check_logged_in() and session['username'] == 'admin':
         #error = "Wrong url"
         select = request.form.get('actions')
         if select == 'read':
@@ -82,12 +102,16 @@ def admin_console():
             return redirect(url_for('update_rest'))
         #print (select)
         return str(select)
+    elif check_logged_in() and session['username'] != 'admin':
+        return str("You don't have privileges to open this page!")
+    else:
+        return str("You are Not logged in")
 
 @app.route('/rest/create', methods=['GET', 'POST'])
 def create_rest():
     error = None
     print (request.method + "#########")
-    if request.method == 'POST':
+    if request.method == 'POST' and check_logged_in() and session['username'] == 'admin':
         print ("yessssookkkkk")
         rest_name = request.form['rest_name']
         print (rest_name)
@@ -102,15 +126,15 @@ def create_rest():
         else:
            error = "Something went wrong"
            return str(error)
-    elif request.method == 'GET':
+    elif request.method == 'GET' and check_logged_in() and session['username'] == 'admin':
         return render_template('create.html')
     else:
-        error = "There is no API with this request"
+        error = "Please login First"
         return str(error)
 @app.route('/rest/delete', methods=['GET', 'POST'])
 def delete_rest():
     error = None
-    if request.method == 'POST':
+    if request.method == 'POST' and check_logged_in() and session['username'] == 'admin':
           rest_name = request.form['rest_name'] 
           result = db_module.delete_(rest_name)
           if result == None:
@@ -119,13 +143,15 @@ def delete_rest():
           else:
               error = "Something went wrong"
               return str(error)
-    else:
+    elif request.method == 'GET' and check_logged_in() and session['username'] == 'admin': 
           return render_template('delete.html')
+    else:
+         return str("Please login first")
 
 @app.route('/rest/update', methods=['GET','POST'])
 def update_rest():        
     error = None
-    if request.method == 'POST':
+    if request.method == 'POST' and check_logged_in() and session['username'] == 'admin':
           rest_name_ = request.form['rest_name']
           category_ = request.form['food_type']
           review_ = request.form['review']
@@ -137,9 +163,10 @@ def update_rest():
           else:
               error = "Something went wrong"
               return str(result)
-    else:
+    elif request.method == 'GET' and check_logged_in() and session['username'] == 'admin':
           return render_template('update.html')
-
+    else:
+          return str("Please Login First")
 
 ##################################################
 ################# Home & Order Pages #############
